requirements.txt
# Local AI Agents Workshop - Dependencies
# Install with: pip install -r requirements.txt

# Core - Anthropic API
anthropic>=0.40.0

# Strands Agents SDK with Anthropic provider
strands-agents[anthropic]>=0.1.0
strands-agents-tools>=0.2.0

# Local RAG (programmatic option with LlamaIndex)
llama-index>=0.11.0
llama-index-llms-anthropic>=0.4.0
llama-index-embeddings-huggingface>=0.3.0
chromadb>=0.5.0

# Local embeddings (free, no API costs)
sentence-transformers>=2.2.0

# Atlassian integration (Jira/Confluence)
atlassian-python-api>=3.41.0

# MCP support
mcp>=1.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Testing
pytest>=7.0.0
